+++ src/kernel/fault.c
@@ -466,7 +466,7 @@ static void double_fault_handler(struct interrupt_frame *frame) {
             "DOUBLE FAULT at RIP=0x%lx CR2=0x%lx cpu=%u pid=%u",
             (unsigned long)frame->rip, (unsigned long)cr2,
             (unsigned int)smp_get_cpu_id(), proc ? (unsigned int)proc->pid : 0);
-        if (n < 0 || n >= (int)sizeof(msg)) {
+        if (n < 0 || (size_t)n >= sizeof(msg)) {
             /* Truncation is acceptable — msg is best-effort here */
         }
         msg[sizeof(msg) - 1] = '\0';
